// frontend/src/shared/api/generated/platform.ts
export interface components {
    schemas: {
// ...
        CollectionTask: {
            /** Collection Task Id */
            collection_task_id: string;
            /** Description */
            description: string;
            /** Name */
            name: string;
            /** Project Id */
            project_id: string;
            /** Quality Threshold */
            quality_threshold: number | null;
            /** Scenario */
            scenario: string;
            /**
             * Schema Version
             * @default 1
             */
            schema_version: string;
            status: components["schemas"]["CollectionTaskStatus"];
            target: components["schemas"]["CollectionTarget"] | null;
            /** Task Code */
            task_code: string;
            /** Type */
            type: string;
        };
        /** CollectionTaskPage */
        CollectionTaskPage: {
            /** Items */
            items: components["schemas"]["CollectionTask"][];
            /** Next Cursor */
            next_cursor?: string | null;
        };
        /** CollectionTaskProgress */
        CollectionTaskProgress: {
            /**
             * As Of
             * Format: date-time
             */
            as_of: string;
            /** Collection Task Id */
            collection_task_id: string;
            observed_sources: components["schemas"]["ManifestObservedSources"];
            /** Project Id */
            project_id: string;
            qc: components["schemas"]["CollectionTaskQcProgress"];
            /** Received Package Count */
            received_package_count: number;
            /**
             * Schema Version
             * @default 1
             */
            schema_version: string;
            status: components["schemas"]["CollectionTaskStatus"];
        };
// ...
        CreateCollectionTask: {
            /**
             * Description
             * @default
             */
            description: string;
            /** Name */
            name: string;
            /** Quality Threshold */
            quality_threshold?: number | null;
            /** Scenario */
            scenario: string;
            target?: components["schemas"]["CollectionTarget"] | null;
            /** Type */
            type: string;
        };
// ...
        UpdateCollectionTask: {
            /** Description */
            description?: string;
            /** Name */
            name?: string;
            /** Quality Threshold */
            quality_threshold?: number | null;
            /** Scenario */
            scenario?: string;
            target?: components["schemas"]["CollectionTarget"] | null;
            /** Type */
            type?: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}